requirements.txt
# AST Visualizer Requirements
flask>=2.3.0
redis>=4.5.0
gitpython>=3.1.0
radon>=6.0.0
psutil>=5.9.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
plotly
networkx
flask-cors

# Multi-language AST support via Tree-sitter
tree-sitter>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0  
tree-sitter-go>=0.20.0
tree-sitter-rust>=0.20.0
tree-sitter-c>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-css>=0.20.0
tree-sitter-html>=0.20.0